+++ src/blocks/header/header 2.css
@@ -1,0 +1,6 @@
+.header {
+  min-height: 132px;
+  display: flex;
+  align-items: center;
+  border-bottom: 1px solid #A5A5A5;
+}
