+++ src/blocks/logo/logo 2.css
@@ -1,0 +1,3 @@
+.logo {
+  width: 148px;
+}
