+++ src/blocks/place-card/__delete-icon/place-card__delete-icon 2.css
@@ -1,0 +1,12 @@
+.place-card__delete-icon {
+  background: url(../../../images/trash-icon.svg) center no-repeat;
+  background-color: transparent;
+  border: none;
+  width: 18px;
+  height: 20px;
+  position: absolute;
+  top: 18px;
+  right: 15px;
+  cursor: pointer;
+  display: none;
+}
