+++ src/blocks/place-card/__description/place-card__description 2.css
@@ -1,0 +1,8 @@
+.place-card__description {
+  display: flex;
+  min-height: 78px;
+  align-items: center;
+  justify-content: space-between;
+  box-sizing: border-box;
+  padding: 0 22px;
+}
