+++ src/blocks/place-card/__image/place-card__image 2.css
@@ -1,0 +1,22 @@
+.place-card__image {
+  height: 282px;
+  background-size: cover;
+  background-position: center;
+  border-top-left-radius: 10px;
+  border-top-right-radius: 10px;
+  position: relative;
+}
+
+.place-card__image:hover {
+  cursor: pointer;
+}
+
+.place-card__image::after {
+  content: '';
+  display: block;
+  width: 282px;
+  height: 282px;
+  background-color: rgba(0, 0, 0, .3);
+  border-top-left-radius: 10px;
+  border-top-right-radius: 10px;
+}
