+++ src/blocks/place-card/__like-icon/place-card__like-icon 2.css
@@ -1,0 +1,17 @@
+.place-card__like-icon {
+  background: url(../../../images/like-inactive.svg) center no-repeat;
+  background-color: transparent;
+  border: none;
+  width: 21px;
+  height: 19px;
+  cursor: pointer;
+  transition: transform .3s linear;
+}
+
+.place-card__like-icon_liked {
+  background-image: url(../../../images/like-active.svg);
+}
+
+.place-card__like-icon:hover {
+  transform: scale(1.03);
+}
