+++ src/blocks/place-card/__like-text/place-card__like-text 2.css
@@ -1,0 +1,5 @@
+.place-card__like-text {
+    margin: 0;
+    font-size: 13px;
+    padding-top: 2px;
+}
