+++ src/blocks/place-card/__like-wrapper/place-card__like-wrapper 2.css
@@ -1,0 +1,6 @@
+.place-card__like-wrapper {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
