+++ src/blocks/place-card/__name/place-card__name 2.css
@@ -1,0 +1,5 @@
+.place-card__name {
+  margin: 0;
+  font-size: 24px;
+  line-height: 29px;
+}
