+++ src/blocks/place-card/place-card 2.css
@@ -1,0 +1,15 @@
+@import url(./__image/place-card__image.css);
+@import url(./__image/_city/_karachaevsk/place-card__image_city_karachaevsk.css);
+@import url(./__delete-icon/place-card__delete-icon.css);
+@import url(./__description/place-card__description.css);
+@import url(./__name/place-card__name.css);
+@import url(./__like-icon/place-card__like-icon.css);
+@import url(./__like-text/place-card__like-text.css);
+@import url(./__like-wrapper/place-card__like-wrapper.css);
+
+.place-card {
+  width: 282px;
+  background-color: #fff;
+  border-radius: 10px;
+  min-height: 360px;
+}
