+++ src/blocks/places-list/places-list 2.css
@@ -1,0 +1,6 @@
+.places-list {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, 282px);
+  grid-gap: 18px 18px;
+  justify-content: center;
+}
