+++ src/blocks/popup/__button/popup__button 2.css
@@ -1,0 +1,12 @@
+.popup__button {
+  width: 100%;
+  height: 50px;
+  border: 1px solid rgba(0, 0, 0, .2);
+  margin-top: 48px;
+  color: rgba(0, 0, 0, .2);
+  font-size: 36px;
+}
+
+.popup__button-save {
+  font-size: 18px;
+}
