+++ src/blocks/popup/__form/popup__form 2.css
@@ -1,0 +1,10 @@
+.popup__form {
+  margin-top: 60px;
+}
+
+
+.popup__form-error {
+  color: red;
+  font-size: 12px;
+  display: none;
+}
