+++ src/blocks/popup/__lightbox/popup__lightbox 2.css
@@ -1,0 +1,35 @@
+@keyframes image {
+    0% {
+        transform: scale(0.1);
+    }
+    80% {
+        transform: scale(1.1);
+    }
+    100% {
+        transform: scale(1);
+    }
+}
+
+.popup__lightbox {
+    display: flex;
+    flex-direction: column;
+    justify-content: center; 
+    align-items: center; 
+    position: relative;
+}
+
+.popup__lightbox-image {
+    display: block; 
+    max-height: 60vh; 
+    max-width: 60vw;
+    border-radius: 15px;
+    animation: image 0.2s ease-in-out;
+}
+
+@media screen and (max-width: 425px) {
+    .popup__lightbox-image {
+        max-width: 100vw;
+            border-radius: 0px;
+    }
+}
+
