+++ src/blocks/popup/__title/popup__title 2.css
@@ -1,0 +1,5 @@
+.popup__title {
+  margin: 0;
+  font-size: 24px;
+  line-height: 30px;
+}
