+++ src/blocks/popup/_is-opened/popup_is-opened 2.css
@@ -1,0 +1,3 @@
+.popup.popup_is-opened {
+  display: flex;
+}
