+++ src/blocks/profile/profile 2.css
@@ -1,0 +1,5 @@
+.profile {
+  min-height: 300px;
+  display: flex;
+  align-items: center;
+}
