+++ src/blocks/root/__section/root__section 2.css
@@ -1,0 +1,10 @@
+.root__section {
+  width: calc(100% - 400px);
+  margin: auto;
+}
+
+@media screen and (max-width: 980px) {
+  .root__section {
+    width: 90%;
+  }
+}
