+++ src/blocks/root/root 2.css
@@ -1,0 +1,16 @@
+@import url(./__section/root__section.css);
+
+.root {
+  background-color: black;
+  min-height: 100vh;
+  font-family: 'Inter';
+  padding-bottom: 194px;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+@media screen and (max-width: 1024px) {
+  .root {
+    padding-bottom: 140px;
+  }
+}
