+++ src/blocks/user-info/__button/user-info__button 2.css
@@ -1,0 +1,12 @@
+.user-info__button {
+  position: absolute;
+  top: 16px;
+  right: 0;
+}
+
+@media screen and (max-width: 980px) {
+  .user-info__button {
+    position: static;
+    margin-bottom: 40px;
+  }
+}
