+++ src/blocks/user-info/__button/user-info__edit-button 2.css
@@ -1,0 +1,8 @@
+.user-info__edit-button {
+    margin-top: 24px;
+    width: 38px;
+    height: 18px;
+    background-color: #000;
+    color: #fff;
+    font-size: 12px;
+}
