+++ src/blocks/user-info/__data/user-info__data 2.css
@@ -1,0 +1,11 @@
+.user-info__data {
+  max-width: 364px;
+  margin-left: 36px;
+}
+
+@media screen and (max-width: 980px) {
+  .user-info__data {
+    margin: 20px 0;
+    text-align: center;
+  }
+}
