+++ src/blocks/user-info/__job/user-info__job 2.css
@@ -1,0 +1,5 @@
+.user-info__job {
+  margin: 20px 0 0 0;
+  font-size: 20px;
+  line-height: 25px;
+}
