+++ src/blocks/user-info/__name/user-info__name 2.css
@@ -1,0 +1,6 @@
+.user-info__name {
+  margin: 0;
+  font-weight: normal;
+  font-size: 54px;
+  line-height: 56px;
+}
