+++ src/blocks/user-info/__photo/user-info__photo 2.css
@@ -1,0 +1,15 @@
+.user-info__photo {
+  width: 150px;
+  height: 150px;
+  min-width: 150px;
+  border-radius: 50%;
+  background-size: cover;
+  background-position: center;
+  background-image: url(../../../images/avatar.jpg);
+}
+
+@media screen and (max-width: 980px) {
+  .user-info__photo {
+    margin-top: 40px;
+  }
+}
