+++ src/blocks/user-info/user-info 2.css
@@ -1,0 +1,21 @@
+@import url(./__photo/user-info__photo.css);
+@import url(./__data/user-info__data.css);
+@import url(./__name/user-info__name.css);
+@import url(./__job/user-info__job.css);
+@import url(./__button/user-info__button.css);
+@import url(./__button/user-info__edit-button.css);
+
+.user-info {
+  display: flex;
+  position: relative;
+  align-items: flex-start;
+  color: white;
+  width: 100%;
+}
+
+@media screen and (max-width: 980px) {
+  .user-info {
+    flex-direction: column;
+    align-items: center;
+  }
+}
