+++ src/index 2.js
@@ -1,0 +1,2 @@
+import "./pages/index.css";
+import "./script/script.js";
